requirements.txt
Django==1.9
MySQL-python==1.2.5
PyJWT==1.4.0
PyYAML==3.11
Pygments==2.0.2
argparse==1.2.1
blessings==1.6
bpython==0.14.2
coverage==4.0.3
curtsies==0.1.21
decorator==4.0.6
django-bower==5.0.4
django-extensions==1.6.1
django-nose==1.4.3
django-rest-swagger==0.3.4
django-widget-tweaks==1.4.1
djangorestframework==3.3.2
greenlet==0.4.9
nose==1.3.7
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyparsing==2.0.7
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.18
traitlets==4.0.0
wsgiref==0.1.2
